+++ artemis-test/src/main/java/com/artemis/component/Size.java
@@ -1,0 +1,7 @@
+package com.artemis.component;
+
+import com.artemis.Component;
+
+public class Size extends Component {
+	public float width, height;
+}