+++ artemis/src/main/java/com/artemis/ArchetypeBuilder.java
@@ -75,7 +75,7 @@ public class ArchetypeBuilder {
 		EntityManager em = world.getEntityManager();
 		return new Archetype(types, em.compositionIdentity(bitset(types)));
 	}
-
+	
 	/** generate bitset mask of types. */
 	private static BitSet bitset(ComponentType[] types) {
 		BitSet bs = new BitSet();
@@ -97,25 +97,5 @@ public class ArchetypeBuilder {
 			types[i] = tf.getTypeFor(classes.get(i));
 		
 		return types;
-	}
-
-	/**
-	 * Provides a blueprint for new entities, offering greatly
-	 * improved insertion performance for systems and managers.
-	 * </p>
-	 * Instance entities using {@link com.artemis.World#createEntity(com.artemis.ArchetypeBuilder.Archetype)}
-	 */
-	public final static class Archetype {
-		final ComponentType[] types;
-		final int compositionId;
-
-		/**
-		 * @param types Desired composition of derived components.
-		 * @param compositionId uniquely identifies component composition.
-		 */
-		public Archetype(ComponentType[] types, int compositionId) {
-			this.types = types;
-			this.compositionId = compositionId;
-		}
 	}
 }