+++ artemis/src/main/java/com/artemis/World.java
@@ -8,13 +8,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-import com.artemis.ArchetypeBuilder.Archetype;
 import com.artemis.annotations.Mapper;
 import com.artemis.annotations.Wire;
 import com.artemis.managers.UuidEntityManager;
 import com.artemis.utils.Bag;
 import com.artemis.utils.ImmutableBag;
 import com.artemis.utils.reflect.ClassReflection;
+import com.artemis.utils.reflect.Constructor;
 import com.artemis.utils.reflect.Field;
 import com.artemis.utils.reflect.ReflectionException;
 
@@ -99,7 +99,9 @@ public class World {
 	private int maxRebuiltIndicesPerTick;
 
 	final EntityEditPool editPool = new EntityEditPool(this);
-
+	
+	private boolean initialized;
+	
 	/**
 	 * Creates a new world.
 	 * <p>
@@ -163,6 +165,7 @@ public class World {
 	 * added.
 	 */
 	public void initialize() {
+		initialized = true;
 		injector.update();
 		for (int i = 0; i < managersBag.size(); i++) {
 			Manager manager = managersBag.get(i);
@@ -187,10 +190,13 @@ public class World {
 	 * @param target Object to inject into.
 	 */
 	public void inject(Object target) {
-		if (injector == null)
-			throw new MundaneWireException("World#initialize() has not been called.");
-		
+		assertInitialized();
 		injector.inject(target);
+	}
+
+	private void assertInitialized() {
+		if (!initialized)
+			throw new MundaneWireException("World#initialize() has not yet been called.");
 	}
 
 	/**
@@ -220,6 +226,24 @@ public class World {
 
 		if (exceptions.size() > 0)
 			throw new ArtemisMultiException(exceptions);
+	}
+	
+	<T extends EntityFactory<T>> T createFactory(Class<T> factory) {
+		if (!factory.isInterface())
+			throw new MundaneWireException("Expected interface for type: " + factory);
+		
+		assertInitialized();
+
+		String impl = factory.getCanonicalName() + "Impl";
+		try {
+			Class<?> implClass = ClassReflection.forName(impl);
+			Constructor constructor = ClassReflection.getConstructor(implClass, World.class);
+			@SuppressWarnings("unchecked")
+			T instance = (T) constructor.newInstance(this);
+			return instance;
+		} catch (ReflectionException e) {
+			throw new RuntimeException(e);
+		}
 	}
 
 	/**
@@ -400,7 +424,7 @@ public class World {
 		added.add(e);
 		return e;
 	}
-
+	
 	/**
 	 * Create and return a new or reused entity instance.
 	 * <p>