+++ artemis/src/main/java/com/artemis/annotations/Sticky.java
@@ -1,0 +1,10 @@
+package com.artemis.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.SOURCE)
+@Target(ElementType.METHOD)
+public @interface Sticky {}