+++ artemis/src/test/java/com/artemis/ArchetypeManagerTest.java
@@ -5,7 +5,6 @@ import static org.junit.Assert.*;
 import org.junit.Before;
 import org.junit.Test;
 
-import com.artemis.ArchetypeBuilder.Archetype;
 import com.artemis.annotations.Wire;
 import com.artemis.component.ComponentX;
 import com.artemis.component.ComponentY;