+++ artemis/src/test/java/com/artemis/ParamArchTest.java
@@ -1,0 +1,22 @@
+package com.artemis;
+
+
+
+public class ParamArchTest {
+	public static class Position extends Component {
+		public float x, y;
+	}
+	
+	public static class Size extends Component {
+		public float width, height;
+	}
+	public static class Asset extends Component {
+		public String path; 
+	}
+	public static class HitPoints extends Component {
+		public int hitpoints; 
+	}
+	public static class Velocity extends Component {
+		public float x, y;
+	}
+}